+++ newclient/src/modules/address/components/TableAddressTokensHeader.vue
@@ -7,89 +7,14 @@
         -->
         <v-layout v-if="isTransfers" align-center justify-start row fill-height>
             <v-flex md3>
-                <v-layout
-                    :class="[!isTransferSortActive(0) && !isTransferSortActive(1) ? 'inactive-sort' : '']"
-                    grid-list-xs
-                    row
-                    align-center
-                    justify-start
-                    fill-height
-                >
-                    <h5 class="pl-3">{{ this.$tc('token.name', 1) }}</h5>
-                    <v-flex d-flex align-center>
-                        <v-layout align-start justify-right>
-                            <v-btn
-                                v-if="!loading && !isTransferSortActive(1)"
-                                :class="[!isTransferSortActive(0) && !isTransferSortActive(1) ? 'inactive-btn' : '']"
-                                flat
-                                icon
-                                @click="selectTransferFilter(0)"
-                            >
-                                <v-icon :class="[isTransferSortActive(0) ? 'white--text' : '']">fas fa-long-arrow-alt-up</v-icon>
-                            </v-btn>
-                            <v-btn v-if="!loading && isTransferSortActive(1)" flat icon @click="selectTransferFilter(1)">
-                                <v-icon :class="[isTransferSortActive(1) ? 'white--text' : '']">fas fa-long-arrow-alt-down</v-icon>
-                            </v-btn>
-                        </v-layout>
-                    </v-flex>
-                </v-layout>
+                <h5 class="pl-3">{{ this.$tc('token.name', 1) }}</h5>
             </v-flex>
             <v-spacer />
             <v-flex md3>
-                <v-layout
-                    :class="[!isTransferSortActive(2) && !isTransferSortActive(3) ? 'inactive-sort' : '']"
-                    grid-list-xs
-                    row
-                    align-center
-                    justify-start
-                    fill-height
-                >
-                    <h5>{{ $t('common.amount') }}</h5>
-                    <v-flex d-flex align-center>
-                        <v-layout align-start justify-right>
-                            <v-btn
-                                v-if="!loading && !isTransferSortActive(2)"
-                                :class="[!isTransferSortActive(3) && !isTransferSortActive(2) ? 'inactive-btn' : '']"
-                                flat
-                                icon
-                                @click="selectTransferFilter(3)"
-                            >
-                                <v-icon :class="[isTransferSortActive(3) ? 'white--text' : '']">fas fa-long-arrow-alt-up</v-icon>
-                            </v-btn>
-                            <v-btn v-if="!loading && isTransferSortActive(2)" flat icon @click="selectTransferFilter(2)">
-                                <v-icon :class="[isTransferSortActive(2) ? 'white--text' : '']">fas fa-long-arrow-alt-down</v-icon>
-                            </v-btn>
-                        </v-layout>
-                    </v-flex>
-                </v-layout>
+                <h5>{{ text }}</h5>
             </v-flex>
             <v-flex md2>
-                <v-layout
-                    :class="[!isTransferSortActive(4) && !isTransferSortActive(5) ? 'inactive-sort' : '']"
-                    grid-list-xs
-                    row
-                    align-center
-                    justify-start
-                    fill-height
-                >
-                    <h5>{{ text }}</h5>
-                    <v-flex d-flex align-center>
-                        <v-layout align-start justify-right>
-                            <v-btn
-                                v-if="!loading && !isTransferSortActive(4)"
-                                :class="[!isTransferSortActive(5) && !isTransferSortActive(4) ? 'inactive-btn' : '']"
-                                flat
-                                icon
-                                @click="selectTransferFilter(5)"
-                            >
-                                <v-icon :class="[isTransferSortActive(5) ? 'white--text' : '']">fas fa-long-arrow-alt-up</v-icon>
-                            </v-btn>
-                            <v-btn v-if="!loading && isTransferSortActive(4)" flat icon @click="selectTransferFilter(4)">
-                                <v-icon :class="[isTransferSortActive(4) ? 'white--text' : '']">fas fa-long-arrow-alt-down</v-icon>
-                            </v-btn>
-                        </v-layout>
-                    </v-flex>
-                </v-layout>
+                <h5>{{ $t('common.age') }}</h5>
             </v-flex>
         </v-layout>
         <!--
@@ -109,7 +34,7 @@
                 >
                     <div class="token-image" />
                     <h5>{{ $t('token.tokenName') }}</h5>
-                    <v-flex d-flex align-center>
+                    <v-flex v-if="hasTokens" d-flex align-center>
                         <v-layout align-start justify-right>
                             <v-btn
                                 v-if="!loading && !isTokenSortActive(1)"
@@ -139,7 +64,7 @@
                     fill-height
                 >
                     <h5 class="pl-1">{{ text }}</h5>
-                    <v-flex d-flex align-center>
+                    <v-flex v-if="hasTokens" d-flex align-center>
                         <v-layout align-start justify-right>
                             <v-btn
                                 v-if="!loading && !isTokenSortActive(3)"
@@ -167,7 +92,7 @@
                     fill-height
                 >
                     <h5>{{ $t('usd.value') }}</h5>
-                    <v-flex d-flex align-center>
+                    <v-flex v-if="hasTokens" d-flex align-center>
                         <v-layout align-start justify-right>
                             <v-btn
                                 v-if="!isTokenSortActive(5)"
@@ -195,7 +120,7 @@
                     fill-height
                 >
                     <h5>{{ $t('token.change') }}</h5>
-                    <v-flex d-flex align-center>
+                    <v-flex v-if="hasTokens" d-flex align-center>
                         <v-layout align-start justify-right>
                             <v-btn
                                 v-if="!loading && !isTokenSortActive(7)"
@@ -219,7 +144,7 @@
 
 <script lang="ts">
 import { Component, Vue, Prop } from 'vue-property-decorator'
-import { TRANSFER_FILTER_VALUES, TOKEN_FILTER_VALUES } from '@app/modules/address/models/AddressSort'
+import { TOKEN_FILTER_VALUES } from '@app/modules/address/models/TokenSort'
 
 @Component
 export default class TableAddressTokensHeader extends Vue {
@@ -230,10 +155,10 @@ export default class TableAddressTokensHeader extends Vue {
    */
     @Prop(Boolean) isErc20!: boolean
     @Prop(Boolean) isTransfers!: boolean
+    @Prop(Boolean) hasTokens!: boolean
     @Prop(Boolean) loading!: boolean
 
     tokenSort = 1
-    transferSort = 1
 
     /*
     ===================================================================================
@@ -256,35 +181,12 @@ export default class TableAddressTokensHeader extends Vue {
         this.$emit('sortBy', TOKEN_FILTER_VALUES[this.tokenSort])
     }
     /**
-     * Select transfer filter and emit to parent
-     * @param _value {Number}
-     */
-    selectTransferFilter(_value: number): void {
-        if (this.isTransferSortActive(_value)) {
-            if (_value % 2 == 0) {
-                _value = _value + 1
-            } else {
-                _value = _value - 1
-            }
-        }
-        this.transferSort = _value
-        this.$emit('sortBy', TRANSFER_FILTER_VALUES[this.transferSort])
-    }
-    /**
      * Check if token filter is active
      * @param _value {Number}
      * @returns {Boolean}
      */
     isTokenSortActive(_value: number): boolean {
         return this.tokenSort === _value
-    }
-    /**
-     * Check if transfer filter is active
-     * @param _value {Number}
-     * @returns {Boolean}
-     */
-    isTransferSortActive(_value: number): boolean {
-        return this.transferSort === _value
     }
 
     /*