+++ newclient/src/modules/address/handlers/AddressTransfers/AddressTransfers.vue
@@ -36,7 +36,7 @@
         <table-txs :max-items="maxItems" :index="index" :is-loading="loading" :table-message="message" :txs-data="transfers" :is-scroll-view="false">
             <template #header>
                 <table-address-txs-header v-if="isETH" :address="address" />
-                <table-address-tokens-header v-else :is-erc20="isERC20" :is-transfers="true" @sortBy="sortTransfers" />
+                <table-address-tokens-header v-else :is-erc20="isERC20" :is-transfers="true"/>
             </template>
             <template #rows>
                 <v-card v-for="(tx, index) in transfers" :key="index" class="transparent" flat>
@@ -78,7 +78,6 @@ import { EthTransfer } from '@app/modules/address/models/EthTransfer'
 import { ErrorMessage } from '../../models/ErrorMessagesForAddress'
 import { getLatestPrices_getLatestPrices as TokenMarketData } from '@app/core/components/mixins/CoinData/apolloTypes/getLatestPrices'
 import { CoinData } from '@app/core/components/mixins/CoinData/CoinData.mixin'
-import { TRANSFER_FILTER_VALUES, TransferSort } from '@app/modules/address/models/AddressSort'
 
 @Component({
     components: {
@@ -117,7 +116,6 @@ import { TRANSFER_FILTER_VALUES, TransferSort } from '@app/modules/address/model
                             this.ethTransfers = data.getEthTransfersV2.transfers.map(item => {
                                 return new EthTransfer(item)
                             })
-                            // this.sortTransfers(TRANSFER_FILTER_VALUES[1])
                         }
                         if (this.initialLoad) {
                             this.showPagination = this.getTransfers.nextKey != null
@@ -274,18 +272,11 @@ export default class AddressTransers extends Mixins(CoinData) {
         return false
     }
 
-    get tokenSort(): TransferSort {
-        return new TransferSort(this.getTransfers, this.isERC20)
-    }
-
     /*
     ===================================================================================
       Methods:
     ===================================================================================
     */
-    sortTransfers(sort: string): void {
-        this.tokenSort.sortTransfers(this.getTransfers, sort)
-    }
     /**
      * Fetches image for the contract
      * @param contract {String}